Added compute shader functionality [app] Now using a temporary compute approach for testing [ogc] Added writeable images ("RenderTarget" with isComputeTarget=true), added compute functionality (computelist). Fixed various compute shader problems. Added isType<T>() in GraphicsObject to avoid dynamic casts [oiSH] Now including storage images in shader reflection [ostlc] Added rgba, stuv specifiers for vector and clamping

+++ ogc/src/graphics/objects/render/drawlist.cpp
@@ -7,7 +7,7 @@ using namespace oi;
 
 const DrawListInfo DrawList::getInfo() { return info; }
 u32 DrawList::getBatches() { return (u32) info.objects.size(); }
-u32 DrawList::getMaxBatches() { return (u32) info.maxBatches; }
+u32 DrawList::getMaxBatches() { return info.maxBatches; }
 GBuffer *DrawList::getBuffer() { return info.drawBuffer; }
 
 void DrawList::flush() {
